requirements.txt
prometheus_client
eternalegypt
python-dotenv